Extract copyImageUrlToClipboard to shared util (#873)
